tests/wpt/tests/scroll-animations/scroll-timelines/scroll-timeline-snapshotting.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ScrollTimeline snapshotting</title>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#event-loop">
<link rel="help" href="https://github.com/whatwg/html/pull/11613">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/web-animations/testcommon.js"></script>
<script src="./testcommon.js"></script>
<script src="/css/css-scroll-snap/support/common.js"></script>

<style>
  body {
    height: 800px;
    width: 800px;
  }
</style>
<div id="log"></div>

<script>
'use strict';

promise_test(async t => {
  const scroller = document.scrollingElement;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;
  const timeline = new ScrollTimeline({source: scroller});
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollTop, 0, "verify test pre-condition");

  scroller.scrollBy({top: 100, behavior: 'smooth'});
  // Wait for the scroll to change.
  await waitForScrollEvent(document);
  // The next frame the scroll changes, the scroll-timelines should not be
  // updated before the scroll events are dispatched. Store the scroll position
  // from the previous frame to compare the timeline to.
  let scroll_percentage = (scroller.scrollTop / maxScroll) * 100;

  await waitForScrollEvent(document);

  assert_percents_equal(
      timeline.currentTime,
      scroll_percentage,
      'Scroll timeline current time corresponds to the scroll position.');
}, 'ScrollTimeline current time is updated after programmatic animated ' +
   'scroll.');

</script>
